+++ src/concretecomponent/Coffee.java
@@ -1,4 +1,8 @@
-public class Coffee extends HotDrinks{
+package concretecomponent;
+
+import component.HotDrinks;
+
+public class Coffee extends HotDrinks {
 
     double price = 0.6;
