+++ src/decorator/Decorator.java
@@ -1,3 +1,7 @@
+package decorator;
+
+import component.HotDrinks;
+
 public abstract class Decorator extends HotDrinks {
 
     public abstract double getPrice();
