+++ src/test/DecoratorTest.java
@@ -1,6 +1,14 @@
+package test;
+
+import component.HotDrinks;
+import concretecomponent.Coffee;
+import concretecomponent.Tea;
+import concretedecorator.Cream;
+import concretedecorator.Sugar;
+
 /**
  * Made by  Adib Ghassani Waluya, Rizki Pramadhan, Yosua Santoso
- * This pattern will decorate the base class HotDrinks with addition of
+ * This pattern will decorate the base class component.HotDrinks with addition of
  * sugar and cream.
  */
 
@@ -8,17 +16,17 @@ public class DecoratorTest {
 
     public static void main(String[] args) {
 
-        //No Decorator
+        //No decorator.Decorator
         HotDrinks blackCoffee = new Coffee();
         HotDrinks tea = new Tea();
 
         System.out.println(blackCoffee.getDescription() + ": EUR " + blackCoffee.getPrice()
                             + "\n" + tea.getDescription() + ": EUR " + tea.getPrice());
 
-        //With Decorator
+        //With decorator.Decorator
 
         HotDrinks coffeeCream = new Cream(new Coffee());
-        HotDrinks coffeeSugar = new Cream(new Coffee());
+        HotDrinks coffeeSugar = new Sugar(new Coffee());
         HotDrinks coffeeComplete = new Cream(new Cream(new Sugar(new Coffee())));
 
         System.out.println(coffeeCream.getDescription() + ": EUR " + coffeeCream.getPrice()
